Compact Reloc from 40 to 32 bytes (u32 target payloads and thunk_off)
Reloc is the highest-count struct in a link - a debug clang link has
~5.9M of them - so its width dominates the cache traffic of every
relocation pass (scan_relocations, thunk layout, apply_relocations).

Shrink two fields that never need 64 bits:

  - RelocTarget::Sym/Section payloads usize -> u32. A target is an
    index into an object's symbol list or the global subsection arena;
    neither approaches 2^32. mold-rust likewise indexes with u32.
  - thunk_off u64 -> u32. It is an offset of a range-extension thunk
    within one output section. Mach-O caps a single section's span far
    below 4 GiB (arm64 __TEXT sits above a 4 GiB __PAGEZERO, and
    ADRP/branch reach is bounded), so u32 is always sufficient; the
    no-thunk sentinel moves from u64::MAX to u32::MAX.

Absolute virtual addresses stay u64: an arm64 executable's addresses
start at 0x1_0000_0000 because of the 4 GiB __PAGEZERO, so they never
fit in 32 bits. 322 integration tests pass.

+++ src/arch/arm64.rs
@@ -366,7 +366,7 @@ impl Arch for Arm64 {
 
             // A relocation refers to either a symbol or a section.
             let (target, addend) = if r.is_extern() {
-                (RelocTarget::Sym(r.r_symbolnum() as usize), addend)
+                (RelocTarget::Sym(r.r_symbolnum() as u32), addend)
             } else {
                 let addr = if r.is_pcrel() {
                     (hdr.addr + r.r_address as u64).wrapping_add_signed(addend)
@@ -379,7 +379,7 @@ impl Arch for Arm64 {
                 else {
                     fatal!(diag, "{file_name}: bad relocation: {}", r.r_address);
                 };
-                let target = RelocTarget::Section(idx);
+                let target = RelocTarget::Section(idx as u32);
                 (target, (addr - sections[idx].addr) as i64)
             };
 
@@ -391,7 +391,7 @@ impl Arch for Arm64 {
                 is_subtracted,
                 target,
                 addend,
-                thunk_off: u64::MAX,
+                thunk_off: u32::MAX,
             });
             i += 1;
         }
@@ -454,12 +454,12 @@ impl Arch for Arm64 {
                     if !(-(1 << 27)..1 << 27).contains(&val) {
                         // Out of reach: branch through the thunk entry
                         // assigned during layout.
-                        if r.thunk_off == u64::MAX {
+                        if r.thunk_off == u32::MAX {
                             error!(ctx, "branch target out of range: {val:x}");
                         } else {
                             let isec = &ctx.isecs[isec_id];
                             let thunk_addr =
-                                ctx.chunks[isec.osec].hdr.addr + r.thunk_off;
+                                ctx.chunks[isec.osec].hdr.addr + r.thunk_off as u64;
                             val = thunk_addr.wrapping_sub(p) as i64;
                         }
                     }
