+++ src/context.rs
@@ -303,7 +303,7 @@ impl<E: Arch> Context<E> {
     /// Returns the symbol a relocation refers to, if it refers to one.
     pub fn reloc_target_sym(&self, obj: usize, rel: &Reloc) -> Option<SymbolId> {
         match rel.target {
-            RelocTarget::Sym(idx) => Some(self.objs[obj].syms[idx]),
+            RelocTarget::Sym(idx) => Some(self.objs[obj].syms[idx as usize]),
             RelocTarget::Section(_) => None,
         }
     }
@@ -311,8 +311,8 @@ impl<E: Arch> Context<E> {
     /// Returns the input section a relocation's target lives in, if any.
     pub fn reloc_target_isec(&self, obj: usize, rel: &Reloc) -> Option<InputSectionId> {
         match rel.target {
-            RelocTarget::Sym(idx) => self.symtab[self.objs[obj].syms[idx]].isec.map(|i| i as usize),
-            RelocTarget::Section(idx) => Some(idx),
+            RelocTarget::Sym(idx) => self.symtab[self.objs[obj].syms[idx as usize]].isec.map(|i| i as usize),
+            RelocTarget::Section(idx) => Some(idx as usize),
         }
     }
 
@@ -329,8 +329,8 @@ impl<E: Arch> Context<E> {
     /// Resolves a relocation target to its output address.
     pub fn reloc_target_addr(&self, obj: usize, rel: &Reloc) -> u64 {
         match rel.target {
-            RelocTarget::Sym(idx) => self.sym_addr(self.objs[obj].syms[idx]),
-            RelocTarget::Section(idx) => self.isec_addr(idx),
+            RelocTarget::Sym(idx) => self.sym_addr(self.objs[obj].syms[idx as usize]),
+            RelocTarget::Section(idx) => self.isec_addr(idx as usize),
         }
     }
 }
